+++ docs/en_US/changelog.md
@@ -1,13 +1,30 @@
+> 2020-05-05
+
+<span style='color:red'>**WARNING :**</span>
+
+<span style='color:red'>**Blink announces for May 11, 2020 the end of support for connections by tools other than IFTTT or their official application.**</span>
+
+<span style='color:red'>**From May 11, the plugin may therefore no longer be usable in Jeedom.**</span>
+
+> 2020-02-28
+  + Fix sort of videos in historical view
+  + Update the file names of downloads from historical view
+
 > 2020-02-26
   + Fix URL to last video and last thumbnail
   + Add battery pourcentage information 
+
 > 2020-02-01 
   + Add URL to last video and last thumbnail
+
 > 2019-10-06
   + Fix camera list (new equipment)
+
 > 2019-09-19
   + Update the configuration management
+
 > 2019-08-08
   + Stable version
+
 > 2019-07-21
   + First beta version