+++ docs/fr_FR/changelog.md
@@ -1,13 +1,30 @@
+> 2020-05-05
+
+<span style='color:red'>**ATTENTION :**</span>
+
+<span style='color:red'>**Blink annonce pour le 11 mai 2020 la fin du support des connexions par des outils autres que IFTTT ou leur application officielle.**
+
+<span style='color:red'>**A partir de cette date, le plugin risque donc de ne plus être utilisable dans Jeedom.**</span>
+
+> 2020-02-28
+  + Correction du tri des vidéos dans la vue historique
+  + Modification du nom des fichiers téléchargés depuis la vue historique
+
 > 2020-02-26
   + Correction des URL de la dernière vidéo et de la vignette
   + Ajout de l'information sur le pourcentage de la batterie (gestion des seuils d'alerte)
+
 > 2020-02-01 
   + Ajout des URL de la dernière vidéo et de la vignette
+
 > 2019-10-06
   + Correction de la liste des caméras (configuration d'un équipement)
+
 > 2019-09-19
   + Modification de la gestion de la configuration
+
 > 2019-08-08
   + Passage en version stable
+
 > 2019-07-21
   + Première version Beta du plugin